Imported using git-ubuntu import.

Publish parent: 60fa76981019c1098ce34022b03b9eecf8098855
Changelog parent: d8d7f71aee604e9b36e4cf88ecc8af13ed952b07
Unapplied parent: 54298af7070cb050fa76e3333155259c2001ad41

+++ sysdeps/mach/hurd/chflags.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 1991-2016 Free Software Foundation, Inc.
+/* Copyright (C) 1991-2015 Free Software Foundation, Inc.
    This file is part of the GNU C Library.
 
    The GNU C Library is free software; you can redistribute it and/or
@@ -24,7 +24,7 @@
 
 /* XXX shouldn't this be __chflags? */
 int
-chflags (const char *file, unsigned long int flags)
+chflags (const char *file, int flags)
 {
   error_t err;
   file_t port = __file_name_lookup (file, 0, 0);
